Vulkan: 4xBRZ texture upscaling shader optimizations

+++ Common/GPU/Vulkan/VulkanProfiler.h
@@ -15,8 +15,8 @@
 
 struct ProfilerScope {
 	std::string name;
-	size_t startQueryId;
-	size_t endQueryId;
+	int startQueryId;
+	int endQueryId;
 	int level;
 };
